this is the commit on day 3

+++ exercises/day-2/HTTPserver.js
@@ -10,30 +10,41 @@ const server = http.createServer((req, res) => {
   res.setHeader('Content-Type','text/plain');
   const parsedUrl = url.parse(req.url, true);
   const pathname = parsedUrl.pathname;
+  const fileName = path.join(directoryPath, parsedUrl?.query?.name??'');
   
-  try {
+  try{
+
+    //home page
+    if(pathname == '//'){
+      res.end('welcome to homepage');
+    }
 
     // List files
-    if (pathname == '/list') {
+    else if (pathname == '/list') {
       fs.readdir(directoryPath, (err, files) => {
         const fileList = files.join('\n');
         res.end(fileList);
       });
-
     }
 
     // Serve file content
     else if (pathname == '/file' && parsedUrl.query.name) {
-      const fileName = path.join(directoryPath, parsedUrl.query.name);
-      fs.readFile(fileName, 'utf8', (err, data) => {
-        res.end(data);
-      });
+      if(parsedUrl.query.name == ''){res.end('incomplete query')}
+      if(parsedUrl.query.name.includes(".txt")){
+        fs.readFile(fileName, 'utf8', (err, data) => {
+          res.end(data);
+        });
+      }
+      res.end('not readable format');
 
     } 
 
     // Create a new file
-    else if (pathname == '/create' && parsedUrl.query.name && parsedUrl.query.content) {
-      const fileName = path.join(directoryPath, parsedUrl.query.name);
+    else if (pathname == '/create' && parsedUrl.query.name && (parsedUrl.query.content??'')) {
+      if(parsedUrl.query.name == ''){res.end('incomplete query')}
+      if(fs.existsSync()){
+        res.end('this file already exists');
+      }
       fs.writeFile(fileName, parsedUrl.query.content, 'utf8', (err) => {
         res.end(`File ${parsedUrl.query.name} created successfully`);
       });
@@ -41,8 +52,8 @@ const server = http.createServer((req, res) => {
     }
     
     // Append content to an existing file
-    else if (pathname == '/append' && parsedUrl.query.name && parsedUrl.query.content) {
-      const fileName = path.join(directoryPath, parsedUrl.query.name);
+    else if (pathname == '/append' && parsedUrl.query.name && (parsedUrl.query.content??'')) {
+      if(parsedUrl.query.name == ''){res.end('incomplete query')}
       fs.appendFile(fileName, `\n${parsedUrl.query.content}`, 'utf8', (err) => {
         res.end(`Content appended to ${parsedUrl.query.name}`);
       });
@@ -51,11 +62,13 @@ const server = http.createServer((req, res) => {
 
     // Delete a file
     else if (pathname == '/delete' && parsedUrl.query.name) {
-      const fileName = path.join(directoryPath, parsedUrl.query.name);
-      fs.unlink(fileName, (err) => {
-        res.end(`File ${parsedUrl.query.name} deleted successfully`);
-      });
-
+      if(parsedUrl.query.name == ''){res.end('incomplete query')}
+      if(fs.existsSync(parsedUrl.query.name)){
+        fs.unlink(fileName, (err) => {
+          res.end(`File ${parsedUrl.query.name} deleted successfully`);
+        });
+      }
+      res.end('no such file exists');
     } 
 
     else {
